# FlightBooking.Admin/bin/Release/net5.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "FlightBooking.Admin",
    "version": "v1"
  },
  "paths": {
...
    "/register": {
      "post": {
        "tags": [
          "Authenticate"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Register"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/login": {
      "post": {
        "tags": [
          "Authenticate"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Login"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Register": {
      "required": [
        "email",
        "fullName",
        "password",
        "userName"
      ],
      "type": "object",
      "properties": {
        "fullName": {
          "type": "string"
        },
        "userName": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
    "Login": {
      "required": [
        "password",
        "userName"
      ],
      "type": "object",
      "properties": {
        "userName": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
...
  }
}