# FlightBooking.User/Models/Register.cs
using System;
using System.Collections.Generic;
using System.ComponentModel.DataAnnotations;
using System.Linq;
using System.Threading.Tasks;

namespace FlightBooking.User.Models
{
    public class Register
    {
        [Required]
        public string FullName { get; set; }
[Required]
        public string  UserName { get; set; }
        [Required]
        public string Email { get; set; }
        [Required]
        public string Password { get; set; }
    }
}